performance test in its own class
